mcp: use approval tooltip ui for tool calls
Closes #277026
