+++ src/vs/workbench/contrib/chat/browser/chatContentParts/toolInvocationParts/chatToolInvocationPart.ts
@@ -94,63 +94,9 @@ export class ChatToolInvocationPart extends Disposable implements IChatContentPa
 			partStore.add(this.subPart.onDidChangeHeight(() => this._onDidChangeHeight.fire()));
 			partStore.add(this.subPart.onNeedsRerender(render));
 
-			// todo@connor4312: Move MCP spinner to left to get consistent auto approval presentation
-			if (this.subPart instanceof ChatInputOutputMarkdownProgressPart) {
-				const approval = this.createApprovalMessage();
-				if (approval) {
-					this.domNode.appendChild(approval);
-				}
-			}
-
 			this._onDidChangeHeight.fire();
 		};
 		render();
-	}
-
-	/** @deprecated Approval should be centrally managed by passing tool invocation ChatProgressContentPart */
-	private get autoApproveMessageContent() {
-		const reason = IChatToolInvocation.executionConfirmedOrDenied(this.toolInvocation);
-		if (!reason || typeof reason === 'boolean') {
-			return;
-		}
-
-		let md: string;
-		switch (reason.type) {
-			case ToolConfirmKind.Setting:
-				md = localize('chat.autoapprove.setting', 'Auto approved by {0}', markdownCommandLink({ title: '`' + reason.id + '`', id: 'workbench.action.openSettings', arguments: [reason.id] }, false));
-				break;
-			case ToolConfirmKind.LmServicePerTool:
-				md = reason.scope === 'session'
-					? localize('chat.autoapprove.lmServicePerTool.session', 'Auto approved for this session')
-					: reason.scope === 'workspace'
-						? localize('chat.autoapprove.lmServicePerTool.workspace', 'Auto approved for this workspace')
-						: localize('chat.autoapprove.lmServicePerTool.profile', 'Auto approved for this profile');
-				md += ' (' + markdownCommandLink({ title: localize('edit', 'Edit'), id: 'workbench.action.chat.editToolApproval', arguments: [reason.scope] }) + ')';
-				break;
-			case ToolConfirmKind.UserAction:
-			case ToolConfirmKind.Denied:
-			case ToolConfirmKind.ConfirmationNotNeeded:
-			default:
-				return;
-		}
-
-
-		return md;
-	}
-
-	/** @deprecated Approval should be centrally managed by passing tool invocation ChatProgressContentPart */
-	private createApprovalMessage(): HTMLElement | undefined {
-		const md = this.autoApproveMessageContent;
-		if (!md) {
-			return undefined;
-		}
-
-		const markdownString = new MarkdownString('_' + md + '_', { isTrusted: true });
-		const result = this.renderer.render(markdownString);
-		this._register(result);
-		result.element.classList.add('chat-tool-approval-message');
-
-		return result.element;
 	}
 
 	private createToolInvocationSubPart(): BaseChatToolInvocationSubPart {
@@ -189,14 +135,12 @@ export class ChatToolInvocationPart extends Disposable implements IChatContentPa
 				ChatInputOutputMarkdownProgressPart,
 				this.toolInvocation,
 				this.context,
-				this.editorPool,
 				this.codeBlockStartIndex,
 				this.toolInvocation.pastTenseMessage ?? this.toolInvocation.invocationMessage,
 				this.toolInvocation.originMessage,
 				resultDetails.input,
 				resultDetails.output,
 				!!resultDetails.isError,
-				this.currentWidthDelegate
 			);
 		}
 
@@ -205,14 +149,12 @@ export class ChatToolInvocationPart extends Disposable implements IChatContentPa
 				ChatInputOutputMarkdownProgressPart,
 				this.toolInvocation,
 				this.context,
-				this.editorPool,
 				this.codeBlockStartIndex,
 				this.toolInvocation.invocationMessage,
 				this.toolInvocation.originMessage,
 				typeof this.toolInvocation.toolSpecificData.rawInput === 'string' ? this.toolInvocation.toolSpecificData.rawInput : JSON.stringify(this.toolInvocation.toolSpecificData.rawInput, null, 2),
 				undefined,
 				false,
-				this.currentWidthDelegate
 			);
 		}
